requirements.txt
django-crontab==0.7.1
djangorestframework==3.14.0
rest-framework-simplejwt==0.0.2
paramiko==3.2.0
psutil==5.9.5